requirements.txt
Flask
requests
openai
python-dotenv
google-cloud-vision
pytest